Refactoring (moving class to another namespace)
