+++ development/actuators/kubernetes-actuator/src/main/java/eubr/atmosphere/tma/actuator/KubernetesActuator.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import eubr.atmosphere.tma.actuator.examples.Configuration;
 import eubr.atmosphere.tma.actuator.services.RestServices;
 
 @RestController
