+++ development/actuators/kubernetes-actuator/src/main/java/eubr/atmosphere/tma/actuator/services/RestServices.java
@@ -11,7 +11,7 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import eubr.atmosphere.tma.actuator.examples.Configuration;
+import eubr.atmosphere.tma.actuator.Configuration;
 
 public class RestServices {
